+++ eelbrain/plot/brain.py
@@ -2,4 +2,4 @@
 # Author: Christian Brodbeck <christianbrodbeck@nyu.edu>
 
 from ._brain import (activation, cluster, dspm, stat, surfer_brain, annot,
-                     bin_table, copy, image, p_map)
+                     bin_table, copy, image, p_map, annot_legend)