+++ doc/source/changes.rst
@@ -7,6 +7,8 @@ Changes
 New in 0.16
 -----------
 
+* New function for plotting a legend for annot-files:
+  :func:`plot.brain.annot_legend`.
 * Epoch definitions in :attr:`MneExperiment.epochs` can now include a ``'base'``
   parameter, which will copy the given "base" epoch and modify it with the
   current definition.